+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.19",
     "@mui/material": "^5.15.2",
     "@react-three/drei": "^9.92.7",
     "@react-three/fiber": "^8.15.13",
