+++ src/App.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import Model from './models/Model'
+import Models from './models/Models'
 import Header from './components/Header'
 
 export default function App() {
   return (
    <>
     <Header/>
-    <Model/>
+    <Models/>
    </>
   )
 }
